+++ packages/client/src/constants/rarities.ts
@@ -1,0 +1,42 @@
+export const getRarities = (tier: number) => {
+  return rarities[tier - 1];
+};
+
+const rarities = [
+  { // 1
+    title: "Legendary",
+    color: "#F6DDCC"
+  },
+  { // 2
+    title: "Legendary",
+    color: "#F6DDCC"
+  },
+  { // 3
+    title: "Legendary",
+    color: "#F6DDCC"
+  },
+  { // 4
+    title: "Legendary",
+    color: "#F6DDCC"
+  },
+  { // 5
+    title: "Epic",
+    color: "#E8DAEF"
+  },
+  { // 6
+    title: "Exotic",
+    color: "#D6EAF8"
+  },
+  { // 7
+    title: "Rare",
+    color: "#D4EFDF"
+  },
+  { // 8
+    title: "Uncommon",
+    color: "#F2F3F4"
+  },
+  { // 9
+    title: "Common",
+    color: "#F2F3F4"
+  }
+];
