+++ packages/client/src/layers/react/components/modals/inventory/ItemGrid.tsx
@@ -45,34 +45,6 @@ const Container = styled.div`
   justify-content: flex-start;
 `;
 
-const Slot = styled.div`
-  position: relative;
-  border: solid black 1.5px;
-  border-radius: 4px;
-
-  margin: 5px;
-
-  align-items: center;
-  justify-content: center;
-  overflow: hidden;
-`;
-
-const Icon = styled.img`
-  height: 50px;
-  width: 50px;
-  padding: 5px;
-`;
-
-const IconClickable = styled.img`
-  height: 50px;
-  width: 50px;
-  padding: .5vw;
-
-  &:hover {
-    background-color: #BBB;
-  }
-`;
-
 const Balance = styled.div` 
   border-top: solid black 1.25px;
   border-left: solid black 1.25px;
