+++ packages/client/src/layers/react/components/modals/lootboxes/Opener.tsx
@@ -19,7 +19,6 @@ interface Props {
 }
 
 export const Opener = (props: Props) => {
-
   const [curBal, setCurBal] = useState(0);
 
   useEffect(() => {
